requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
openai>=1.0.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
matplotlib>=3.7.0
seaborn>=0.12.0
joblib>=1.3.0
tqdm>=4.65.0
openpyxl>=3.1.2  # For Excel file support
python-Levenshtein>=0.21.1
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
isort>=5.12.0
mypy>=1.5.1
flake8>=6.1.0
torch>=2.0.0  # Required for sentence-transformers
transformers>=4.30.0  # Required for sentence-transformers
huggingface-hub>=0.15.0  # Required for sentence-transformers
pydantic>=2.0.0  # Required for openai
httpx>=0.24.0  # Required for openai
python-dateutil>=2.8.2  # Required for pandas
pytz>=2023.3  # Required for pandas
scipy>=1.10.0  # Required for scikit-learn
